add edit and delete access control

+++ studybud/base/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, redirect
 from django.db.models import Q
+from django.http import HttpResponse
 from django.contrib.auth.decorators import login_required
 from django.contrib.auth.models import User
 from django.contrib import messages
@@ -61,6 +62,7 @@ def room(request, pk):
   context = {'room': room}
   return render(request, 'base/room.html', context)
 
+@login_required(login_url='login')
 def create_room(request):
   form = RoomForm()
   if request.method == 'POST':
@@ -72,9 +74,14 @@ def create_room(request):
   context = {'form': form}
   return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='login')
 def update_room(request, pk):
   room = Room.objects.get(id=pk)
   form = RoomForm(instance=room)
+
+  if request.user != room.host:
+    return HttpResponse('You are not the owner of this room.')
+
   if request.method == 'POST':
     form = RoomForm(request.POST, instance=room)
     if form.is_valid():
@@ -83,8 +90,11 @@ def update_room(request, pk):
   context = {'form': form}
   return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='login')
 def delete_room(request, pk):
   room = Room.objects.get(id=pk)
+  if request.user != room.host:
+    return HttpResponse('You are not the owner of this room.')
   if request.method == 'POST':
     room.delete()
     return redirect('home')
